+++ CHANGELOG.md
@@ -9,7 +9,11 @@
 - Extra details of a finished scannerjob now holds information in regard to objects
   skipped by last modified check. 
 
-  
+### General improvements
+
+- Labels in Outlook are no longer named "OS2datascanner X", but rather
+  "OSdatascanner X". Already existing labels are not changed.
+
 ## Version 3.23.1, 25th June 2024
 
 "Testing, Testing, Is This Thing On?"
@@ -225,8 +229,6 @@
 
 - Existing ScheduledCheckups are no longer duplicated when a setting on a
   scanner is changed, followed by a scan.
-  
-### General improvements
 
 - The selected rule on scannerjobs no longer resets to default value on re-render.